Cleaned up trees
